requirements.txt
python-dotenv==1.0.0
telethon==1.36.0

